+++ controllers/k6_controller.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 
 	"github.com/go-logr/logr"
-	"github.com/grafana/k6-operator/api/v1alpha1"
+	"github.com/grafana/k6-operator/apis/k6/v1alpha1"
 	batchv1 "k8s.io/api/batch/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
