+++ pkg/resources/jobs/helpers_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/go-test/deep"
-	"github.com/grafana/k6-operator/api/v1alpha1"
+	"github.com/grafana/k6-operator/apis/k6/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 )
