+++ pkg/types/types.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/grafana/k6-operator/api/v1alpha1"
+	"github.com/grafana/k6-operator/apis/k6/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 )
